moving all py files to src directory adding functions to filter blast results and find genes via prodigal
